Change md mixin to support better min/max
This commit modifies the `md` mixin defined in `mixins.scss` to accept as parameters both a `min` and a `max` width in a non exclusive way.
Before we were forced to always specify a `min` width while the `max` was optional. Now they are both optional.
Note that using `md` without `min` or `max` will cause an error, but we can specify only `min`, `max`, or both.