Add `Navbar.svelte` component
This commit adds the main navbar component of the application. Since the navbar needs to display icons, this commit also introduces the `Icon.svelte` component to help display a icon with the correct size.

In this commit we also add all the svg icons needed by the application. The icons are located under `src/lib/assets/icons`.
To load svg icons as svelte components we use the vite preprocessing plugin `@poppanator/sveltekit-svg` added as dependency in this commit.