Add cards, headers and project page
This commit adds part of the content of the projects page. For this reason in this commit we introduce the `SectionHeader.svelte` component and the `ProjectCard.svelte` component.

This commit also adds the `BaseCard.svelte` component that is used by the ProjectCard and will be used by other cards in the future.