Change `Icon.svelte` to load the icon by its name
This commit changes the component `Icon.svelte` to load his icon dynamically specifying only the icon name. Before we where locading each icon as a svelte component and the Icon was only an helper for styling it. Now it is the responsible for loading the correct icon. This way it is possible to have elements with a configurable icon.