Redesign the footer component
In this commit we introduce a slight redesign of the footer component. Since all the constats are moved to the contats page there is no need for including them also in the footer. For this reason the footer has been redesigned to have a copyright text just like before and a funny sundial at te bottom.