Fix problem displaying num of papers & coauthors
This commit fixes an issue where the number of papers and co-authors where displayed incorrectly. The bug was caused by a previous change where the structure of the `publicstion.json` file was change to group the papers by year, but the code responsible of counting the papers and the co-authors was not updated correctly.