Add publication-list page
This commit adds a sub-page to the publications page that displays all the papers in an easy to copy format. This is useful in the case we need a list of all the papers so we ca simply copy and paste it whereever we need without messing with the format of the text.